+++ src/alsyntaxmodifiers/dataClassificationModifier.ts
@@ -1,18 +1,14 @@
 import * as vscode from 'vscode';
 import { DevToolsExtensionContext } from "../devToolsExtensionContext";
-import { TextEditorHelper } from '../tools/textEditorHelper';
 import { NumberHelper } from '../tools/numberHelper';
-import { ToolsWorkspaceCommandRequest } from '../langserver/toolsWorkspaceCommandRequest';
-import { SyntaxModifier } from './syntaxModifier';
 import { ToolsWorkspaceCommandResponse } from '../langserver/toolsWorkspaceCommandResponse';
+import { WorkspaceCommandSyntaxModifier } from './workspaceCommandSyntaxModifier';
 
-export class DataClassificationModifier extends SyntaxModifier {
+export class DataClassificationModifier extends WorkspaceCommandSyntaxModifier {
     protected _dataClassification : string | undefined;
 
     constructor(context: DevToolsExtensionContext) {
-        super(context, "addDataClassification");
-        this._showProgress = true;
-        this._progressMessage = "Processing project files. Please wait...";
+        super(context, "Add Data Classification", "addDataClassification");
         this._dataClassification = undefined;
     }
 
@@ -29,21 +25,19 @@ export class DataClassificationModifier extends SyntaxModifier {
         return (confirmation === 'Yes');
     }
 
-    protected showWorkspaceSuccessMessage(response: ToolsWorkspaceCommandResponse) {
-        vscode.window.showInformationMessage(
-            NumberHelper.zeroIfNotDef(response.parameters.noOfChanges).toString() +                    
+    protected getSuccessWorkspaceMessage(response: ToolsWorkspaceCommandResponse): string {
+        return NumberHelper.zeroIfNotDef(response.parameters.noOfChanges).toString() +                    
             ' data classification(s) added to ' +
             NumberHelper.zeroIfNotDef(response.parameters.noOfChangedFiles).toString() +
-            ' file(s).');
+            ' file(s).';
     }
 
-    protected showDocumentSuccessMessage(response: ToolsWorkspaceCommandResponse) {
-        vscode.window.showInformationMessage(
-            response.parameters.noOfChanges.toString() + 
-            ' data classification(s) added.');
+    protected getSuccessDocumentMessage(response: ToolsWorkspaceCommandResponse): string {
+        return response.parameters.noOfChanges.toString() + 
+            ' data classification(s) added.';
     }
 
-    protected async askForParameters() {
+    async askForParameters(uri: vscode.Uri | undefined): Promise<boolean> {
         let valuesList = ['AccountData', 'CustomerContent',
             'EndUserIdentifiableInformation', 'EndUserPseudonymousIdentifiers',
             'OrganizationIdentifiableInformation', 'SystemMetadata'];
@@ -53,6 +47,8 @@ export class DataClassificationModifier extends SyntaxModifier {
             canPickMany: false,
             placeHolder: 'Select Data Classification'
         });
+
+        return true;
     }
 
 }