+++ src/alsyntaxmodifiers/fieldCaptionsModifier.ts
@@ -1,13 +1,11 @@
 import * as vscode from 'vscode';
 import { DevToolsExtensionContext } from "../devToolsExtensionContext";
-import { SyntaxModifier } from "./syntaxModifier";
+import { WorkspaceCommandSyntaxModifier } from './workspaceCommandSyntaxModifier';
 
-export class FieldCaptionsModifier extends SyntaxModifier {
+export class FieldCaptionsModifier extends WorkspaceCommandSyntaxModifier {
     
     constructor(context: DevToolsExtensionContext) {
-        super(context, "addFieldCaptions");
-        this._showProgress = true;
-        this._progressMessage = "Processing project files. Please wait...";
+        super(context, "Add Table Field Captions", "addFieldCaptions");
     }
 
     protected getParameters(uri: vscode.Uri): any {