+++ src/alsyntaxmodifiers/fixIdentifiersCaseModifier.ts
@@ -1,12 +1,10 @@
 import { DevToolsExtensionContext } from "../devToolsExtensionContext";
-import { SyntaxModifier } from "./syntaxModifier";
+import { WorkspaceCommandSyntaxModifier } from "./workspaceCommandSyntaxModifier";
 
-export class FixIdentifiersCaseModifier extends SyntaxModifier {
+export class FixIdentifiersCaseModifier extends WorkspaceCommandSyntaxModifier {
 
     constructor(context: DevToolsExtensionContext) {
-        super(context, "fixIdentifiersCase");
-        this._showProgress = true;
-        this._progressMessage = "Processing project files. Please wait...";
+        super(context, "Fix Identifiers Case", "fixIdentifiersCase");
     }