+++ src/alsyntaxmodifiers/iSntaxModifierResult.ts
@@ -1,0 +1,5 @@
+export interface ISyntaxModifierResult {
+    success: boolean;
+    message: string | undefined;
+    source: string | undefined;
+}