+++ src/alsyntaxmodifiers/iSyntaxModifierFactoriesCollection.ts
@@ -1,0 +1,5 @@
+import { SyntaxModifier } from "./syntaxModifier";
+
+export interface ISyntaxModifierFactoriesCollection {
+    [name: string]: () => SyntaxModifier;
+}