+++ src/alsyntaxmodifiers/sortVariablesModifier.ts
@@ -1,12 +1,10 @@
 import { DevToolsExtensionContext } from "../devToolsExtensionContext";
-import { SyntaxModifier } from "./syntaxModifier";
+import { WorkspaceCommandSyntaxModifier } from "./workspaceCommandSyntaxModifier";
 
-export class SortVariablesModifier  extends SyntaxModifier {
+export class SortVariablesModifier  extends WorkspaceCommandSyntaxModifier {
 
     constructor(context: DevToolsExtensionContext) {
-        super(context, "sortVariables");
-        this._showProgress = true;
-        this._progressMessage = "Processing project files. Please wait...";
+        super(context, "Sort Variables", "sortVariables");
     }
 
 }