+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
 ## 3.0.24
-  - Issue 301 - PermissionSet wizard - PermissionSet name length cannot be longer than 20 characters
+  - Issue #301 - PermissionSet wizard - PermissionSet name length cannot be longer than 20 characters
+  - Issue #295 - New "Code Cleanup" commands for workspace and editor
   
 ## 3.0.24
  - Issue #293 - Sorting code-actions as commands for current file and active project